requirements.txt
fastapi
openai-clip
torchvision
torch